Lot's of refactoring
